# Track and report damage to compatible compositors