--- a/alacritty/src/display.rs
+++ b/alacritty/src/display.rs
@@ -14,6 +14,7 @@
 
 //! The display subsystem including window management, font rasterization, and
 //! GPU drawing.
+use std::cmp::min;
 use std::cmp::max;
 use std::f64;
 use std::fmt;
@@ -21,6 +22,7 @@ use std::time::Instant;
 
 use glutin::dpi::{PhysicalPosition, PhysicalSize};
 use glutin::event_loop::EventLoop;
+use glutin::Rect;
 use log::{debug, info};
 use parking_lot::MutexGuard;
 
@@ -122,6 +124,9 @@ pub struct Display {
     renderer: QuadRenderer,
     glyph_cache: GlyphCache,
     meter: Meter,
+
+    damage_supported: bool,
+    fully_damaged: bool,
 }
 
 impl Display {
@@ -228,6 +233,8 @@ impl Display {
             _ => (),
         }
 
+        let damage_supported = window.swap_buffers_with_damage_supported();
+
         Ok(Display {
             window,
             renderer,
@@ -235,6 +242,8 @@ impl Display {
             meter: Meter::new(),
             size_info,
             font_size: config.font.size,
+            damage_supported,
+            fully_damaged: false,
         })
     }
 
@@ -349,6 +358,7 @@ impl Display {
             PhysicalSize::new(f64::from(self.size_info.width), f64::from(self.size_info.height));
         self.renderer.resize(&self.size_info);
         self.window.resize(physical);
+        self.fully_damaged = true;
     }
 
     /// Draw the screen
@@ -358,7 +368,7 @@ impl Display {
     /// This call may block if vsync is enabled
     pub fn draw<T>(
         &mut self,
-        terminal: MutexGuard<'_, Term<T>>,
+        mut terminal: MutexGuard<'_, Term<T>>,
         message_buffer: &MessageBuffer,
         config: &Config,
     ) {
@@ -368,6 +378,52 @@ impl Display {
         let metrics = self.glyph_cache.font_metrics();
         let glyph_cache = &mut self.glyph_cache;
         let size_info = self.size_info;
+
+        // Check grid damage
+        let damage: Option<Vec<Rect>> = if self.damage_supported {
+            let (width, height, cell_width, cell_height, padding_x, padding_y) = size_info.to_u32();
+
+            if self.fully_damaged {
+                // We need to fully damaged, so let's clear damage and stop
+                // here.
+                terminal.get_damage();
+                self.fully_damaged = false;
+                Some(vec![Rect { x: 0, y: 0, width, height }])
+            } else {
+                // Fetch and clear damage
+                Some(terminal.get_damage()
+                    .iter()
+                    .map(|d| {
+                        // Make end coordinates be lower right corner instead of upper
+                        // left. Also add one more to end_x to compensate for the
+                        // possibility of the end position being a double width
+                        // character.
+                        let (x, y, end_x, end_y) =
+                            (d.x as u32, d.y as u32, (d.end_x + 2) as u32, (d.end_y + 1) as u32);
+
+                        // Then, convert the grid to a rect in gl coordinates
+                        let rect = Rect {
+                            x: x * cell_width + padding_x,
+                            y: height - end_y * cell_height - padding_y,
+                            width: (end_x - x) * cell_width,
+                            height: (end_y - y) * cell_height,
+                        };
+
+                        // And finally, add half a cell of horizontal, quarter of a
+                        // cell of vertical padding to cover overdraw.
+                        let x = rect.x.saturating_sub(cell_width / 2);
+                        let y = rect.y.saturating_sub(cell_height / 4);
+                        Rect {
+                            x: x,
+                            y: y,
+                            width: min(width-x, rect.width + cell_width),
+                            height: min(height-y, rect.height + cell_height / 2),
+                        }
+                    }).collect())
+            }
+        } else {
+            None
+        };
 
         // Update IME position
         #[cfg(not(windows))]
@@ -454,8 +510,11 @@ impl Display {
                 api.render_string(&timing[..], size_info.lines() - 2, glyph_cache, Some(color));
             });
         }
-
-        self.window.swap_buffers();
+        if let Some(damage) = damage {
+            self.window.swap_buffers_with_damage(&damage);
+        } else {
+            self.window.swap_buffers();
+        }
     }
 }