--- a/alacritty/src/display.rs
+++ b/alacritty/src/display.rs
@@ -20,6 +20,7 @@ use std::fmt;
 use std::time::Instant;
 
 use glutin::dpi::{PhysicalPosition, PhysicalSize};
+use glutin::event::{Event as GlutinEvent};
 use glutin::event_loop::EventLoop;
 use glutin::Rect;
 use log::{debug, info};
@@ -370,20 +371,27 @@ impl Display {
         mut terminal: MutexGuard<'_, Term<T>>,
         message_buffer: &MessageBuffer,
         config: &Config,
+        event_queue: &mut Vec<GlutinEvent<Event>>,
     ) {
         let grid_cells: Vec<RenderableCell> = terminal.renderable_cells(config).collect();
+        let visual_bell_animating = terminal.visual_bell.animating();
         let visual_bell_intensity = terminal.visual_bell.intensity();
         let background_color = terminal.background_color();
         let metrics = self.glyph_cache.font_metrics();
         let glyph_cache = &mut self.glyph_cache;
         let size_info = self.size_info;
 
+        // Request immediate re-draw if visual bell animation is not finished
+        if visual_bell_animating {
+            event_queue.push(GlutinEvent::UserEvent(Event::Wakeup));
+        }
+
         // Check grid damage
         let damage: Option<Vec<Rect>> = if self.damage_supported {
             let (width, height, cell_width, cell_height, padding_x, padding_y) =
                 size_info.into_u32();
 
-            if self.fully_damaged {
+            if self.fully_damaged || visual_bell_animating {
                 // We need to fully damaged, so let's clear damage and stop
                 // here.
                 terminal.clear_damage();