Implement damage tracking and reporting.

This allows compositors to only process damaged (that is, updated) regions of our window buffer, which for larger window sizes (think 4k) should significantly reduce compositing workload under compositors that support/honor it, which is good for performance and battery life.

On Wayland, clients are expected to report always report correct damage, so this makes us a good (or at least, better) citizen there. It can also aid remote desktop (waypipe, RDP, VNC, ...) and other types of screencopy by having damage bubble up correctly.

Damage tracking has been implemented in a somewhat naïve fashion, but I believe it is a fair start.

Note that this currently depends on a fork of glutin where support has been added.

Fixes #3186.