@chrisduerr 2019-08-16
I'd be curious how this performs with the whole screen updating, where the damage tracking code is just dead weight.

Have you run any benchmarks on this?