@kennylevinsen 2019-08-16
Not yet—I just finalized this initial revision, and thought I'd publish it early for comments.

It might be worth having some heuristic to bail fine-grained damage tracking (there is also no benefit to reporting absurd amounts of damage rects), but there will probably always be a slight overhead in tracking what cells are damaged. This information can, however, later be reused to minimize redraw (it is already used now to bail on undamaged grids).