@kennylevinsen 2019-08-16
https://github.com/rust-windowing/glutin/pull/1207 opened for the glutin side to make room for discussion of the underlying API.