@chrisduerr 2019-08-16
> We'd also have to figure out how to consider reduced compositor workload/increased efficiency when we get there.

I disagree with this. If the performance gets worse, it's worse. Everything after that is mostly an excuse and not a sufficient justification. Of course it's good when the compositor needs to do less work, but if it turns out that's the more efficient approach, then that's that.