@kchibisov 2019-08-17
I've just ran it on my system sway + 27inch 4k monitor and the input delay is amazing. Browsing mails in mutt is super responsive now, same for scrolling and navigating in vim. I use high key repeat rate, so the cursor "inertia" was a bit noticeable before, but on this PR there is no more such problems. Great!!!

Some performance measurements that should be relevant to this PR.
env:
`tput cols: 320`
`tput lines: 90`

bench:
```
vtebench -w (tput cols) -h (tput lines) -c -b 500000000 alt-screen-random-write > test
cat test
perf stat -r 5 cat test
```
PR: `6.89314 +- 0.00912 seconds time elapsed  ( +-  0.13% )`
master: `6.4391 +- 0.0102 seconds time elapsed  ( +-  0.16% )`

I've taken `alt-screen-random-write`, since it's probably the most relevant for this change. The PR is in WIP state, so I've decided not to benchmark it "hard".