@kchibisov 2019-08-17
@kennylevinsen Are you aware of presented rendering issues (you're forgetting to send damage in some cases)? The PR is in a WIP state, so that's why I'm asking.