@kchibisov 2019-08-17
@chrisduerr FYI the improvement in latency is **very** noticeable on my system.