@kennylevinsen 2019-08-17
> I disagree with this. If the performance gets worse, it's worse. Everything after that is mostly an excuse and not a sufficient justification. Of course it's good when the compositor needs to do less work, but if it turns out that's the more efficient approach, then that's that.

@crisduerr I must have misworded myself if it came across as an excuse for introducing bad performance.

What I am saying is basically just that performance is more than synthetic draw() execution time, and I don't have an overview over real-world alacritty performance testing.

For example, an hypothetical average real-world 5% increase in runtime for draw may be paired with an average real-world 99% decrease in compositor runtime, leading to fewer cycles-to-screen and improved latency. This is a win for Alacritty, even though it is technically doing more work. As long as we're not impairing other platforms, that is.

@kchibisov I thought I had finally gotten it all tracked, but I just noticed a glitch, so I guess I'm not done. I made this branch my daily driver to try to shake out areas that might be missed.

I get your bench to be a 7% regression in vtebench.

The implementation is pretty darn naïve right now: It is a separate walk over the grid from normal rendition, and it will create damage rects for arbitrarily complex damage (including, say, every second cell changed). We should be able to get the number down to a more manageable size.

Also, not included with this is using damage to power partial rendition, which should be able to help further.