@kchibisov 2019-08-17
@kennylevinsen Just a few very noticeable glitches:

- You're not updating cursor properly. Try opening two terminal instances and switching focus between them with `unfocused_hollow: true`. You'll probably see about what I'm talking
- Selection is not properly cleared. Just select a region and then click outside the selection region, it should be cleared after click.

I'll take a closer look on your implementation later.