@kchibisov 2019-08-17
I've just left a few "generic" comments about current implementation. However the main problem with that RP I have is its current structure. I think we can improve it.

From my point of view, the "damage" can be handled like `underline`/`strikeout`, because it's just a flag on a cell, so we can "shape" damage rects (by line) and the using this bunch of damaged lines in `swap_buffers_with_damage` like we're doing it here for ["lines"](https://github.com/jwilm/alacritty/blob/d9d698614ce828ba96ae101f280b37c5c80f607b/alacritty_terminal/src/display.rs#L524).

Did you investigate on integrating dirty handling into `RendarableCell/RendarableCellsIter` stack though?