@kchibisov 2019-08-17
@kennylevinsen one more issue is that you're probably off by `1` with damage regions somewhere. Look at the gap on the screenshot. The selection handling is wrong at the moment, but it reflects another issue I have with chars that occupy entire cell in `neomutt` (continuous lines in threads). You can try to repro it by triple click on the current line, then triple click on one line bellow and the triple click on previous line again.

![Screenshot_Sat 17 Aug 2019 06:29:59 PM MSK](https://user-images.githubusercontent.com/27620401/63214167-743edd00-c11d-11e9-97bb-3f3857611eaf.png)