@kennylevinsen 2019-08-17
> From my point of view, the "damage" can be handled like underline/strikeout, because it's just a flag on a cell, so we can "shape" damage rects (by line) and the using this bunch of damaged lines in swap_buffers_with_damage like we're doing it here for "lines".

> Did you investigate on integrating dirty handling into RendarableCell/RendarableCellsIter stack though?

So there are two things here: Integrating with the main iteration over the grid (which has obvious benefits), and integrating it into that particular loop.

These are separate, as RenderableCellsIter only yields current non-background cells, which is not a superset of damaged cells.

However, I see two solutions to that:
1. Change RenderableCellsIter to yield either all cells, or at least also all dirty cells.
2. Integrate the damage tracking logic as state into RenderableCellsIter, and simply extract it with a call on the iterator afterwards.

Integrating it like so was actually my first idea. However, as I'm still not super good friends with the borrow checker, I have had to go through a few iterations to get a *working* implementation. :)

However, doing this does eliminate one of the small wins of dirty tracking: We won't be able to bail rendition with no dirty cells before partial rendition becomes possible.

> one more issue is that you're probably off by 1 with damage regions somewhere.

I am not at all surprised. :)

Looking at the code, I see that I will probably need to track last selection like I track last cursor in order to damage the old selection properly.