@kchibisov 2019-08-17
@kennylevinsen FYI information alacritty doesn't render empty cells at all, so if cell is empty just ignore it entirely. `alacritty` has very interesting opacity due to this behavior, which usually can't be achieved by tweaking your apps. e.g. this is `alsamixer` which "sets" background color to black, but it's transparent because my `colors.primary.background` equals to `foreground.normal.black` (If you make them different alacritty will render `alsamixer` opaque).

![Screenshot_Sat 17 Aug 2019 07:05:11 PM MSK](https://user-images.githubusercontent.com/27620401/63214553-2678a380-c122-11e9-9f79-e5258958d8ec.png)