@kchibisov 2019-10-24
One more benchmark iteration.

```
This PR with damage tracking(Wayland):
  alt: 11.5739 +- 0.0274 seconds time elapsed  ( +-  0.24% )
  scrolling: 2.5377 +- 0.0141 seconds time elapsed  ( +-  0.55% )
master(Wayland):
  alt: 11.30900 +- 0.00447 seconds time elapsed  ( +-  0.04% )
  scrolling: 2.7114 +- 0.0481 seconds time elapsed  ( +-  1.77% )
This PR without damage tracking(X11):
  alt: 11.2964 +- 0.0770 seconds time elapsed  ( +-  0.68% )
  scrolling: 2.7293 +- 0.0374 seconds time elapsed  ( +-  1.37% )
master(X11):
  alt: 11.05196 +- 0.00433 seconds time elapsed  ( +-  0.04% )
  scrolling: 2.6328 +- 0.0564 seconds time elapsed  ( +-  2.14% )
```

I don't like X11 performance gap though, since they're paying for nothing, however I've noticed that when damage is not supported we're not lowering our damage tracking overhead in `Term`, so we can infinitely push to our damage_list(seems like it) in some cases for example. I think passing `damage_support` to a `Term::new()` or setting its status later, and adding a "guard" to mostly all `self.damage_list.push()` should make sense, just `self.damage = ...` are probably fine here. However we should definitely guard update_damage call.

P.s.
Interesting how the gap is similar in percent to a Wayland one.